% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R, R/test.R
\name{echo}
\alias{echo}
\alias{echo}
\title{Echo}
\usage{
echo(echo)

echo(echo)
}
\arguments{
\item{echo}{A word or sentence to echo}

\item{echo}{A word or sentence to echo}
}
\description{
This function echos whatever you give it.

This function echos whatever you give it.
}
\examples{

echo('This is a test')

echo('This is a test')
}
