% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr.R
\name{pkgr}
\alias{pkgr}
\title{Install Packages on a new workstation}
\usage{
pkgr(category = c("tidyverse", "spatial", "visual", "system",
  "development", "statistics", "data", "other"), dependencies = TRUE)
}
\description{
Spatial offers packages for working with spatial data     Visual offers packages for visualizing data, including ggplot and supporting libraries
}
